refactor: 쿠폰 모달 coupon feature로 이동
